modification ThreadsServer

+++ Projet/src/aquarium/ElementImage.java
@@ -4,12 +4,12 @@ import java.awt.Image;
 import java.awt.Toolkit;
 
 public class ElementImage {
-	private int idClient;
+	private long idClient;
 	private String imageS; // supposé être temporaire. On doit trouver comment charger l'image pour chaque poisson.
 	private Image image;
 	private String nom;
 	
-	public ElementImage(int idC, String img, String n){
+	public ElementImage(long idC, String img, String n){
 		idClient = idC;
 		imageS = img;
 		image = Toolkit.getDefaultToolkit().createImage(ClassLoader.getSystemResource(img));
@@ -17,7 +17,7 @@ public class ElementImage {
 		nom = n;
 	} 
 	
-	public int getidClient(){
+	public long getidClient(){
 		return idClient;
 	}
